re-add provides/conflicts/replaces
With these in the control file, removing the man-db dummy and installing
the real one will result in apt trying to remove both mandoc and the
ragnarok-base packages. This seems like a problem, so these lines were
removed, but keeping them is actually the right way to go. Ragnarok uses
mandoc, plain and simple.

The only time man-db is really needed is when building packages, and I
only build packages in a chroot (miniroot.tgz) or virtual machines,
*never* on my main system, so installing man-db there is not a problem
(miniroot lacks mandoc and other base packages because they are not
needed). If I ever decide to allow other people to build packages, they
will have to follow that guideline of always use the chroot or a VM to
build packages.